feat: set Gin mode to release and log to file in production

+++ controllers/user.go
@@ -56,11 +56,12 @@ func (uc *UserController) Register(c *gin.Context) {
 
 func (uc *UserController) UpdateUserPassword(c *gin.Context) {
 	var req dto.UpdatePasswordRequest
+
 	if err := c.ShouldBindJSON(&req); err != nil {
 		resp.RespHelper.Fail(c, resp.Error.Code, err.Error())
 		return
 	}
-	user_id := c.MustGet("userID").(uint64)
+	user_id := c.MustGet("userID").(uint)
 	err := uc.UserService.UpdateUserPassword(req, user_id)
 	if err != nil {
 		resp.RespHelper.Fail(c, resp.Error.Code, err.Error())
